% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krr_fit.R
\name{krr_fit}
\alias{krr_fit}
\title{Kernel Ridge Regression fit}
\usage{
krr_fit(X, y, rho = 1, lambda = 1e-04)
}
\arguments{
\item{X}{Matrix or data frame of predictors.}

\item{y}{Response vector.}

\item{rho}{Kernel bandwidth parameter.}

\item{lambda}{Penalty parameter.}
}
\value{
A list containing:
  \item{coefficients}{Estimated dual coefficients}
  \item{X}{Training data}
  \item{rho}{Kernel bandwidth}
  \item{lambda}{Penalty parameter}
}
\description{
Fits a Kernel Ridge Regression (KRR) model using a Gaussian kernel.
}
