% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krr_predict.R
\name{predict.krr}
\alias{predict.krr}
\title{Predict method for KRR}
\usage{
\method{predict}{krr}(model, X_new, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{A fitted KRR model from krr_fit().}

\item{newdata}{New data matrix.}
}
\value{
Predicted response vector.
}
\description{
Predicts new responses using a fitted KRR model.
}
